Fix Texture usage of Color
Color was previously a struct. By creating an array of Colors
(i.e. Color[,] C), each Color were assigned their default value
(i.e. 0.0f for R,G and B).

Color was changed to a class by following .NET guidelines :
* https://msdn.microsoft.com/en-us/library/ms229017%28v=vs.110%29.aspx
* "avoid defining a struct unless it has all the following
  characteristics: [...] the type is immutable."

Since Color is now a class, the array of Colors contains
non-instantiated Colors when it is created. In order for this to
work, From255(...) is now a constructor and is used in Textures.