ZBuffer Renderer
At the beginning, the Scene knew how to draw itself, and called
the Canvas to draw each pixel. The Canvas used a ZBuffer to
resolve the visibility problem.

The previous commit moved the ZBuffer usage in the Scene, thus
the Canvas was more simple (only used to draw pixels).

But still, the Scene only knew how to draw itself with a ZBuffer.
It was a problem since we also want to implement Raycasting and
Raytracing. This commit introduces the concept of Renderer. A
Renderer renders a Scene on a Canvas. The Scene does not know
how to draw itself anymore, this task has been delegated to
renderers. The ZBuffer is now a Renderer.

A small performance tweak has also been done. We previously
computed the illumination and then checked if the current object
was visible. This was done in this order due to the Canvas
performing the ZBuffer check. But doing it in this order obviously
had an impact on performances. Now, we only compute illumination
if the object is visible.