fix shadow ray
old: check if there is an intersection with at least one object, and if it
is the case there is a shadow.

problem: not all intersections are valid and we discard some of them (i.e. if
the intersection is behind the light source). If the intersection is discarded
and there was a closer intersection, no shadow will be displayed even though it
should be.

new: retrieve the closest intersection